# EmployeeApi/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Emplyee API",
    "description": "Employee CRUD",
    "version": "v1"
  },
  "paths": {
    "/employeeservice/getemployees": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Gets a list of Employees",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employeeservice/getemployee/{id}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Gets an employee by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employeeservice/updateemployee/{id}": {
      "put": {
        "tags": [
          "Employee"
        ],
        "summary": "Update an employee",
        "parameters": [
          {
            "name": "empId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employeeservice/postemployee": {
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "POST an employee details",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Employees"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employeeservice/delete/{id}": {
      "delete": {
        "tags": [
          "Employee"
        ],
        "summary": "Delete an employees",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmployeeAddress": {
        "type": "object",
        "properties": {
          "addressId": {
            "type": "string",
            "nullable": true
          },
          "address1": {
            "type": "string",
            "nullable": true
          },
          "address2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zipcode": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "employees": {
            "$ref": "#/components/schemas/Employees"
          }
        },
        "additionalProperties": false
      },
      "Employees": {
        "type": "object",
        "properties": {
          "empoyeeId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "emailId": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/EmployeeAddress"
          }
        },
        "additionalProperties": false
      }
    }
  }
}